improvements

+++ python/cudf/cudf/core/column/temporal_base.py
@@ -306,11 +306,11 @@ class TemporalBaseColumn(ColumnBase, Scannable):
             to_res, _ = np.datetime_data(to_dtype)
             max_val = self.max()
             if isinstance(max_val, (pd.Timedelta, pd.Timestamp)):
-                max_val = max_val.to_numpy()
+                max_val = max_val.to_numpy().astype(self.dtype)
             max_val = max_val.astype(self._UNDERLYING_DTYPE, copy=False)
             min_val = self.min()
             if isinstance(min_val, (pd.Timedelta, pd.Timestamp)):
-                min_val = min_val.to_numpy()
+                min_val = min_val.to_numpy().astype(self.dtype)
             min_val = min_val.astype(self._UNDERLYING_DTYPE, copy=False)
             # call-overload must be ignored because numpy stubs only accept literal strings
             # for time units (e.g., "ns", "us") to allow compile-time validation,
